chore: start refactor the project structure
