+++ packages/cli/src/cli.ts
@@ -2,14 +2,14 @@
 
 import pkg from "@pkg";
 import { getRenderCommand } from "@/commands/render";
-import { getCommand } from "@/services/command";
-import { getIssuesCollector } from "@/services/issue";
+import { getCommand } from "@/libs/command";
+import { getIssuesCollector } from "@/libs/issue";
 
 const description = `Dotsecret is designed to simplify the process of fetching secrets from secrets managers and render files with them.
 
 It uses configuration and templates files to specify the secrets to fetch and the output files to render. Ideal for automating the creation of .env, .npmrc, and other files you don't want on your git repository.`;
 
-const run = async () => {
+const cli = async () => {
   const command = getCommand({
     name: "dotsecret",
     summary: pkg.description,
@@ -25,7 +25,7 @@ const run = async () => {
   await command.parseAsync(process.argv);
 };
 
-run().catch((error) => {
+cli().catch((error) => {
   const issuesCollector = getIssuesCollector().addError(error);
   issuesCollector.print({ severity: "error" });
   process.exit(1);
