+++ packages/cli/src/vendors/.eslintrc.cjs
