+++ packages/cli/src/vendors/chalk.ts
