+++ packages/cli/src/vendors/color-json.ts
